@@ -10,19 +10,19 @@
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-static": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.45.1",
 		"mdsvex": "^0.11.0",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"rehype-autolink-headings": "^6.1.1",
 		"rehype-slug": "^5.1.0",
-		"svelte": "^4.2.7",
-		"svelte-preprocess": "^5.0.3",
-		"vite": "^5.0.12"
+		"svelte": "^5.0.0",
+		"svelte-preprocess": "^6.0.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module"
 }
