@@ -3,7 +3,13 @@
 	import HamburgerSVG from './svg/HamburgerSVG.svelte';
 	import XSVG from './svg/XSVG.svelte';
 
-	export let closeOnly = false;
+	/**
+	 * @typedef {Object} Props
+	 * @property {boolean} [closeOnly]
+	 */
+
+	/** @type {Props} */
+	let { closeOnly = false } = $props();
 
 	const toggleIsMenuOpen = () => {
 		isMenuOpen.set(!$isMenuOpen);
@@ -11,7 +17,7 @@
 </script>
 
 <button
-	on:click={toggleIsMenuOpen}
+	onclick={toggleIsMenuOpen}
 	aria-pressed={$isMenuOpen}
 	class="menu-button"
 	tabindex={$isMenuOpen || !closeOnly ? '0' : '-1'}
