@@ -1,12 +1,18 @@
 <script>
 	import { postsPerPage } from '$lib/config'
 
-	export let currentPage
-	export let totalPosts
-	export let path = '/blog/page'
+	/**
+	 * @typedef {Object} Props
+	 * @property {any} currentPage
+	 * @property {any} totalPosts
+	 * @property {string} [path]
+	 */
+
+	/** @type {Props} */
+	let { currentPage, totalPosts, path = '/blog/page' } = $props();
 	
-	let pagesAvailable
-	$: pagesAvailable = Math.ceil(totalPosts / postsPerPage)
+	let pagesAvailable = $derived(Math.ceil(totalPosts / postsPerPage))
+	
 
 	const isCurrentPage = (page) => page == currentPage
 </script>
