@@ -3,7 +3,7 @@
 	import Pagination from '$lib/components/Pagination.svelte'
 	import { siteDescription } from '$lib/config'
 
-	export let data
+	let { data } = $props();
 </script>
 
 
