@@ -4,12 +4,12 @@
 	import Pagination from '$lib/components/Pagination.svelte'
   import { postsPerPage } from '$lib/config'
 
-	export let data
+	let { data } = $props();
 
   const { page, posts, category, total } = data
 
-	$: lowerBound = (page * postsPerPage) - (postsPerPage - 1) || 1
-	$: upperBound = Math.min(page * postsPerPage, total)
+	let lowerBound = $derived((page * postsPerPage) - (postsPerPage - 1) || 1)
+	let upperBound = $derived(Math.min(page * postsPerPage, total))
 </script>
 
 
