@@ -1,4 +1,6 @@
 <script>
+	import { preventDefault } from 'svelte/legacy';
+
 	import MainNav from './MainNav.svelte'
 	import HamburgerMenuButton from './HamburgerMenuButton.svelte'
 	import { siteTitle } from '$lib/config'
@@ -11,7 +13,7 @@
 
 
 <header>
-	<a on:click|preventDefault={focusMain} class="skip-to-content-link" href="#main">
+	<a onclick={preventDefault(focusMain)} class="skip-to-content-link" href="#main">
 		Skip to main content
 	</a>
 	
