@@ -1,4 +1,7 @@
 <script>
+	import { createBubbler, preventDefault } from 'svelte/legacy';
+
+	const bubble = createBubbler();
 	import Callout from '$lib/components/Callout.svelte'
 </script>
 
@@ -16,7 +19,7 @@ If you're using this starter for your own site, feel free to delete this page, o
 
 <Callout>This form does nothing! It's just here to show default styling.</Callout>
 
-<form on:submit|preventDefault>
+<form onsubmit={preventDefault(bubble('submit'))}>
 	<div class="form-section">
 		<label for="name">Name</label>
 		<input type="text" id="name" placeholder="First name" />
