@@ -1,13 +1,13 @@
-import { sveltekit } from '@sveltejs/kit/vite'
+import { sveltekit } from "@sveltejs/kit/vite";
 
 /** @type {import('vite').UserConfig} */
 const config = {
 	plugins: [sveltekit()],
 	server: {
 		fs: {
-			allow: ['.']
-		}
-	}
+			allow: ["."],
+		},
+	},
 };
 
-export default config
+export default config;
